Add setup for manual tests and remove testing.go from acceptance pkg
